refactored as per comment https://github.com/TheCSharpAcademy/CodeReviews.Console.ExerciseTracker/pull/21#issuecomment-1904698942
