+++ ExerciseTracker.StevieTV/ExerciseTracker.StevieTV.sln
@@ -1,0 +1,15 @@
+﻿Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExerciseTracker.StevieTV", "ExerciseTracker.StevieTV\ExerciseTracker.StevieTV.csproj", "{60406FE4-50C8-41EB-BB74-BDEB641C7C9F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{60406FE4-50C8-41EB-BB74-BDEB641C7C9F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{60406FE4-50C8-41EB-BB74-BDEB641C7C9F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{60406FE4-50C8-41EB-BB74-BDEB641C7C9F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{60406FE4-50C8-41EB-BB74-BDEB641C7C9F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
